chore(metrics): move HandlerMethodDecorator commons
